+++ src/Pages/Accounts/Components/AccountCharts/Components/AccountVestingTable.tsx
@@ -10,19 +10,36 @@ export const AccountVestingTable = ({
   isLoading: boolean;
 }) => {
   const nextVesting = data.periodicVestingList.find((element) => !element.isVested);
+  // Calculate sum of vested tokens
+  let currentlyVested = 0;
+  // Because we want to break as soon as we find an item not vested, use a for loop
+  const lengthOfVestingSched = Number(data.periodicVestingList.length);
+  const periodicVestingSched = data.periodicVestingList;
+  if (periodicVestingSched) {
+    for (let i = 0; i < lengthOfVestingSched; i++) {
+      if (periodicVestingSched[i].isVested) {
+        // Add all currently vested amounts
+        currentlyVested += Number(periodicVestingSched[i].coins[0].amount);
+      } else {
+        break;
+      }
+    }
+  }
   const tableData = [
     {
-      param_name: 'Start Date',
+      param_name: 'Start',
       value: new Date(data.startTime).toLocaleDateString(),
     },
     {
-      param_name: 'End Date',
+      param_name: 'End',
       value: new Date(data.endTime).toLocaleDateString(),
     },
     {
       param_name: 'Next Amount',
       value: nextVesting
-        ? formatDenom(Number(nextVesting.coins[0].amount), nextVesting.coins[0].denom)
+        ? formatDenom(Number(nextVesting.coins[0].amount), nextVesting.coins[0].denom, {
+            decimal: 2,
+          })
         : 'N/A',
     },
     {
@@ -30,11 +47,15 @@ export const AccountVestingTable = ({
       value: nextVesting ? new Date(nextVesting.vestingDate).toLocaleDateString() : 'Fully Vested',
     },
     {
-      param_name: 'Amount Granted',
+      param_name: 'Granted',
       value: formatDenom(
         Number(data.originalVestingList[0].amount),
         data.originalVestingList[0].denom
       ),
+    },
+    {
+      param_name: 'Vested',
+      value: formatDenom(Number(currentlyVested), data.originalVestingList[0].denom),
     },
   ];
 
@@ -44,7 +65,7 @@ export const AccountVestingTable = ({
       tableData={tableData}
       tableHeaders={[
         {
-          displayName: 'Vesting Info',
+          displayName: 'Info',
           dataName: 'param_name',
         },
         {